IMPORTANT! Major changes to various config files.
In an attempt to clean up my $HOME from random dotfiles, I'm moving
things over to the corresponding XDG base directories. This, however,
can break stuff if it's not done properly. It works fine on Foxes, but
replicating the move on Hekate will have to be done carefully.

Things to keep in mind:

1) .xinitrc in $HOME is now a link to $XDG_CONFIG_HOME/X11/xinitrc, because
startx seems to be hardcoded to look for $HOME/.xinitrc, while xinit
does find the file in the proper path. Since I use startx, I kind of
need to keep the init file there.

2) For Neovim, I dropped Vundle in favour of vim-plug. It should be
   pretty much the same thing, but the plugins will have to be
   reinstalled once.

3) ZSH is being broken into multiple (2 for now) files to keep the
   config manageable. zaliases exists now in $ZDOTDIR.

In other news, I had the brilliant idea mixing in other, unrelated
changes to Qtile, kitty, dunst, etc. Most of those are just switching
the system font to JetBrains Mono.